Move error storage out of `PluginConfiguration`
It was too easy to accidentally keep things between runs.
