+++ plugins/formal-verification/formver.common/src/org/jetbrains/kotlin/formver/PluginConfiguration.kt
@@ -11,31 +11,9 @@ class PluginConfiguration(
     val conversionSelection: TargetsSelection,
     val verificationSelection: TargetsSelection,
 ) {
-    private val internalErrorInfos = mutableListOf<String>()
-    private val minorErrors = mutableListOf<String>()
-
     init {
         if (conversionSelection < verificationSelection) {
             throw IllegalArgumentException("Conversion options may not be stricter than verification options; converting $conversionSelection but verifying $verificationSelection.")
         }
-    }
-
-    fun clearMinorErrors() {
-        minorErrors.clear()
-    }
-
-    fun addMinorError(error: String) {
-        minorErrors.add(error)
-    }
-
-    fun forEachMinorError(action: (String) -> Unit) {
-        minorErrors.forEach(action)
-    }
-
-    fun formatErrorWithInfos(error: String): String =
-        internalErrorInfos.joinToString(prefix = "$error\n", separator = "\n")
-
-    fun addErrorInfo(msg: String) {
-        internalErrorInfos.add(msg)
     }
 }
